==> notebooks/0_README.ipynb <==
<h1> # Essai de webscrapping de trustpilot 2023 10 01 LH</h1>

<h3> Objectif </h3>
collecter les avis et notes des utilisateurs concernant une catégories d'entreprises


<h3> Structure du site </h3>
page principale
    catégories
        liste des entreprises notées
            page des entreprises avec impression des avis

<h3> Stratégie de scrappping </h3>

1) Choisir une catégorie 
2) Scrap 1 : depuis la page des listes de entreprises:
    scrapper:
- nom de l'entreprise
- trust score
- nombre d'avis
- lien vers la page de l'entreprise

<u> -> Sauvegarder comme df/csv "liste_entreprises" </u>

3) Scrap 2:
Utiliser la liste des entreprises pour scrapper sur chaque page entreprise:
Avis:
- Nom utilisateur/ice
- Nombre d'avis
- Localisation
- Date de l'avis
- Titre de l'avis
- Contenu de l'avis
- Date de l'expérience

<u>-> Sauvegarder en DF/CSV "avis"</u>

! Il faut réussir à le faire sur toutes les pages à partir du lien:
ex: https://fr.trustpilot.com/review/liberkeys.com?page=155
Comment savoir le nombre total de pages?
- en essayant d'extraire la page jusuq'à erreur 404
- ou en scrappant le bouton qui affiche la dernière page?

Renforcer les métadonnées entreprise:
scrapper:
- A propos de [entreprise]
- email
- telephone
- adresse
- positionnement par rapport aux autres entreprises de la catégorie

<u> -> Renforcer DF/CSV "liste_entreprises" </u>
