==> notebooks/3_compilation_df.ipynb <==
Ce Script identifie et merge les csv de scrapping pour les différentes pages d'entreprises.
Le tout est sauvegardé sous data/avis/general_df.csv

v 20 octobre 2023 LH 

import pandas as pd
import os

groot = "../data/avis/csv_entreprises/"  

# Get the list of CSV files
list_csv = os.listdir(groot+ "")
list_csv_clean = [csv for csv in list_csv if "csv" in csv]

# Initialize an empty dataframe for consolidation
general_df = pd.DataFrame()

# Loop through each CSV and append to the general dataframe
for csv_key in list_csv_clean:
    print(csv_key)
    print(groot+csv_key)
    
    # Attempt reading with Python engine, and if there's an error, skip bad lines
    try:
        temp_df = pd.read_csv(groot + str(csv_key), engine='python')
    except:
        temp_df = pd.read_csv(groot + str(csv_key), error_bad_lines=False)
    
    # Append the temporary dataframe to the consolidated dataframe
    general_df = pd.concat([general_df, temp_df])

# Print the number of records in the consolidated dataframe
print("nombre d'avis dans la df consolidée:", len(general_df))

# sauvegarde de la df consolidée
general_df.to_csv("../data/avis/"+"general_df.csv")
