==> notebooks/4_code_dataclean.ipynb <==
### Import Lib

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import locale
import dateparser

### Import CSV

df = pd.read_csv("../data/avis/general_df.csv", index_col=0)
df = df.drop(columns=["Unnamed: 0","user"], axis=1)
df.head()

# on drop YOUNITEDCREDIT trop d'avis positifs/bots
df = df[df.page != "YounitedCredit" ]
df.head()
df.page.value_counts()

### Cleaning

df_cleaned = df.dropna()
df_cleaned = df_cleaned.drop(columns=["localisation","user"])
df_cleaned = df_cleaned.rename(columns={"page":"Société"})
df_cleaned.info()

 Transformation étoiles en Int64 de 1 à 5

df_cleaned["etoiles"] = pd.to_numeric(df_cleaned["etoiles"], errors="coerce")
df_cleaned = df_cleaned[df_cleaned["etoiles"].between(1, 5)] # Vérifier que les valeurs soient bien 1 à 5
df_cleaned["etoiles"] = df_cleaned["etoiles"].astype("int64")

df_cleaned.head()
df_cleaned.info()

Transformation titre et text en STR

df_cleaned["titre_avis"] = df_cleaned["titre_avis"].astype(str)
df_cleaned["text_avis"] = df_cleaned["text_avis"].astype(str)
df_cleaned['text_total'] = df_cleaned.titre_avis + " ; " + df_cleaned.text_avis
df_cleaned = df_cleaned.drop(["titre_avis","text_avis"], axis =1)
df_cleaned.head()

Transformation date_experience en datetime avec un dictionnaire

df_cleaned["date_experience"] = df_cleaned["date_experience"].str.replace("Date de l'expérience:", "").str.strip()

months_fr_to_en = {
    'janvier': 'January',
    'février': 'February',
    'mars': 'March',
    'avril': 'April',
    'mai': 'May',
    'juin': 'June',
    'juillet': 'July',
    'août': 'August',
    'septembre': 'September',
    'octobre': 'October',
    'novembre': 'November',
    'décembre': 'December'
}

def replace_month_fr_to_en(date_str):
    for fr, en in months_fr_to_en.items():
        date_str = date_str.replace(fr, en)
    return date_str

df_cleaned["date_experience"] = df_cleaned["date_experience"].apply(replace_month_fr_to_en)
df_cleaned["date_experience"] = pd.to_datetime(df_cleaned["date_experience"])

print(df_cleaned["date_experience"][0:2])

df_cleaned.head()

Transformation date_avis en datetime

# date_experience, date_avis > datetime
from datetime import datetime
df_cleaned['date_avis'] = pd.to_datetime(df_cleaned['date_avis'])

def format_date(x):
    if pd.isnull(x):
        return x  
    return x.strftime('%Y-%m-%d %H:%M:%S')

df_cleaned['date_avis'] = df_cleaned['date_avis'].apply(format_date)
df_cleaned['date_avis'] = pd.to_datetime(df_cleaned['date_avis'])
df_cleaned.head()

df_cleaned.info()

Save cleaned csv

df_cleaned.to_csv("../data/avis/df_cleaned.csv", encoding="utf-8", index=False)
