==> notebooks/6_Sentiment_Camembert.ipynb <==
from transformers import pipeline,AutoTokenizer, TFCamembertForSequenceClassification
import pandas as pd
from tqdm import tqdm
import numpy as np

df = pd.read_csv('../data/avis/df_clean_noYC_lemma.csv')
df.head()

import sentencepiece
# chargement du modèle
tokenizer = AutoTokenizer.from_pretrained("tblard/tf-allocine")
model = TFCamembertForSequenceClassification.from_pretrained("tblard/tf-allocine")

classifier=  pipeline("text-classification", model = model, tokenizer= tokenizer)
#tokenizer_kwargs = {'padding':True,'truncation':True,'max_length':512,'return_tensors':'pt'}

indice = 5
outputs = classifier(df['text_total'][indice])
print(outputs, df['text_total'][indice])

# slow but more reliable than batching

MAX_LENGTH = 512  # nombre de mots max pour Camembert

sentiment = pd.DataFrame()
df['label'] =np.nan
df['score'] = np.nan
for i, text in tqdm(enumerate(df.text_total), total=df.shape[0]):
    # Tronquer le texte s'il est trop long
    text = text[:MAX_LENGTH]
    try:
        temp = classifier(text)
        temp = pd.DataFrame(temp)
        df.loc[i,'label'] = temp.label.iloc[0]
        df.loc[i,'score'] = temp.score.iloc[0]
    except Exception as e:
        print(f"Erreur lors du traitement du texte à l'indice {i}. Erreur: {e}")

# transformation des scores pour intégrer la polarité et la confiance du modèle
# vers -1 est très probablement négativ, vers +1 est très probablement positif
df['label'] = df.label.replace("NEGATIVE", -1)
df['label'] = df.label.replace("POSITIVE", 1)
df['sentiment_norm'] = df["label"] * df["score"]
df.head()

df.to_csv("../data/avis/gen_clean_lemma_sent_noYc_.csv")
