==> notebooks/7_creation_df_equilibre_modalites.ipynb <==
import pandas as pd
df = pd.read_csv('../data/avis/gen_clean_lemma_sent_noYc_.csv',
                  sep=",", index_col = 0)
df = df.join(pd.get_dummies(df.Société, prefix = 'bank_'))
df.head(2)

# Mélange aléatoire des lignes du jeu de données
df = df.sample(frac = 1, random_state=7)
# split entre jeu de test et d'entrainement
train_ratio = 0.8
total_rows = df.shape[0]
train_size = int(total_rows*train_ratio)
train = df[:train_size]
test =  df[train_size:]

print("total rows:",total_rows,"\n","train size:", train_size, 
      "\n","test size:",total_rows - train_size)


# subsampling manuel sur la base de la modalité minoritaire


Train

# normalizing
from sklearn.preprocessing import StandardScaler

# Insérez votre code ici
norm_train = train[['longueur_texte', "n_avis"]]
scaler = StandardScaler().fit(norm_train) 
train[['longueur_texte', "n_avis"]] = scaler.transform(train[['longueur_texte', "n_avis"]])



mod_etoiles = train.etoiles.value_counts().sort_values()
n_mod = mod_etoiles.min()
train1 = train[train.etoiles ==1].iloc[0:n_mod,:]
train2 = train[train.etoiles ==2].iloc[0:n_mod,:]
train3 = train[train.etoiles ==3].iloc[0:n_mod,:]
train4 = train[train.etoiles ==4].iloc[0:n_mod,:]
train5 = train[train.etoiles ==5].iloc[0:n_mod,:]

train_equal = pd.concat([train1,train2,train3,train4,train5])

# On mélange à nouveau 
train_equal = train_equal.sample(frac = 1, random_state=7)
train_equal.etoiles.value_counts()

# saving balanced train dataset
train_equal.to_csv("../data/avis/train_noYC_lemma_sent_equil.csv")

Test

test[['longueur_texte', "n_avis"]] = scaler.transform(test[['longueur_texte', "n_avis"]])

mod_etoiles = test.etoiles.value_counts().sort_values()
n_mod = mod_etoiles.min()
test1 = test[test.etoiles ==1].iloc[0:n_mod,:]
test2 = test[test.etoiles ==2].iloc[0:n_mod,:]
test3 = test[test.etoiles ==3].iloc[0:n_mod,:]
test4 = test[test.etoiles ==4].iloc[0:n_mod,:]
test5 = test[test.etoiles ==5].iloc[0:n_mod,:]

test_equal = pd.concat([test1,test2,test3,test4,test5])

# On mélange à nouveau 
test_equal = test_equal.sample(frac = 1, random_state=7)
test_equal.etoiles.value_counts()

# saving balanced test dataset
test_equal.to_csv("../data/avis/test_noYC_lemma_sent_equil.csv")
