==> notebooks/8_Z_Benchmark_Results.ipynb <==
# compiling Benchmark results
from os import listdir
import pandas as pd
import time

bench_list = listdir("../reports/benchmark")
bench_list.remove(".DS_Store")

benchmark = pd.DataFrame()
for file in bench_list:
    df = pd.read_csv(f'../reports/benchmark/{file}', index_col=0)
    benchmark = pd.concat([benchmark, df])

new_cols = ['model', 'grid search', 'score', 'precision',
       'recall', 'f1', 'time_taken_mns', 'run_date', 'used/best params']
benchmark = benchmark.reindex(columns=new_cols)
benchmark = benchmark.sort_values("f1", ascending = False)
benchmark

benchmark.to_csv(f"../reports/classification_benchmark_{time.strftime('%Y-%m-%d', time.localtime())}.csv")
