==> notebooks/10_chargement_modèles_prediction_streamlit.ipynb <==
# Chargement du modèle Camembert pour la prédiction d'étoiles

from transformers import CamembertTokenizer, CamembertForSequenceClassification
import torch

model = CamembertForSequenceClassification.from_pretrained('../models/camembert/model_dst_camembert/', num_labels=5)
tokenizer = CamembertTokenizer.from_pretrained('../models/camembert/tokenizer_dst_camembert')

model.eval()

# Texte d'exemple
text = "J'adore ce produit !"
inputs = tokenizer(text, return_tensors="pt", padding=True, truncation=True, max_length=512)

with torch.no_grad():
    outputs = model(**inputs)
    probabilities = torch.nn.functional.softmax(outputs.logits, dim=-1)
    predicted_class = torch.argmax(probabilities, dim=-1)

print("Probabilities:", probabilities.numpy())
print("Predicted class:", predicted_class.item() +1) # on rajoute 1 car classes d'entrainement de 0 à 4
